+++ Kernel/GenericInterface/Operation/Ticket/TicketGet.pm
@@ -376,9 +376,14 @@ sub Run {
             next TICKET;
         }
 
+        my $ArticleTypes;
+        if ( $UserType eq 'Customer' ) {
+            $ArticleTypes = [ $TicketObject->ArticleTypeList( Type => 'Customer' ) ];
+        }
         my @ArticleBoxRaw = $TicketObject->ArticleGet(
             TicketID          => $TicketID,
             ArticleSenderType => $ArticleSenderType,
+            ArticleType       => $ArticleTypes,
             DynamicFields     => $DynamicFields,
             Extended          => $Extended,
             Order             => $ArticleOrder,